added CSV/PSV export option
activated automatically in cm controllers that have an export_sql_query
and an export_list_fields attribute, using the CSV option. CSV is comma
separated and PSV is pipe separated.

also fixed a bug whereby </td>'s and </tr>'s where not being printed,
except on the last row of a list!

A bad one, I know. and it's been outstanding for months. '